Added variable reset angles. Cleanup. Updated readme.

+++ src/detection/ObjectDetector.h
@@ -16,8 +16,8 @@ namespace Detect {
 		bool found;
 		int label;
 		std::string target;
-		cv::Point2d center;
-		cv::Rect2d boundingBox;
+		cv::Point2i center;
+		cv::Rect2i boundingBox;
 	};
 	class ObjectDetector
 	{
